Move materialization of dirs for Nailgun server to NailgunPool
