+++ src/rust/engine/process_execution/src/nailgun/mod.rs
@@ -1,6 +1,6 @@
 use std::collections::btree_map::BTreeMap;
 use std::collections::btree_set::BTreeSet;
-use std::os::unix::fs::symlink;
+//use std::os::unix::fs::symlink;
 use std::path::{Path, PathBuf};
 use std::sync::Arc;
 use std::time::Duration;
@@ -8,7 +8,7 @@ use std::time::Duration;
 use boxfuture::{try_future, BoxFuture, Boxable};
 use futures::future::Future;
 use futures::stream::Stream;
-use hashing::Digest;
+//use hashing::Digest;
 use log::{debug, trace};
 use nails::execution::{child_channel, ChildInput, ChildOutput, Command};
 use tokio::net::TcpStream;
@@ -32,10 +32,10 @@ mod parsed_jvm_command_lines_tests;
 use async_semaphore::AsyncSemaphore;
 pub use nailgun_pool::NailgunPool;
 use parsed_jvm_command_lines::ParsedJVMCommandLines;
-use std::fs::{read_link, remove_file};
+//use std::fs::{read_link, remove_file};
 use std::net::SocketAddr;
-use store::Store;
-use workunit_store::WorkUnitStore;
+//use store::Store;
+//use workunit_store::WorkUnitStore;
 
 // Hardcoded constants for connecting to nailgun
 static NAILGUN_MAIN_CLASS: &str = "com.martiansoftware.nailgun.NGServer";
@@ -161,50 +161,6 @@ impl CommandRunner {
   fn calculate_nailgun_name(main_class: &str) -> NailgunProcessName {
     format!("nailgun_server_{}", main_class)
   }
-
-  // TODO(#8481) When we correctly set the input_files field of the nailgun EPR, we won't need to pass it here as an argument.
-  // TODO(#8489) We should move this code to NailgunPool. This returns a Future, so this will involve making the struct Futures-aware.
-  fn materialize_workdir_for_server(
-    store: Store,
-    workdir_for_server: PathBuf,
-    requested_jdk_home: PathBuf,
-    input_files: Digest,
-    workunit_store: WorkUnitStore,
-  ) -> BoxFuture<(), String> {
-    // Materialize the directory for running the nailgun server, if we need to.
-    let workdir_for_server2 = workdir_for_server.clone();
-
-    // TODO(#8481) This materializes the input files in the client req, which is a superset of the files we need (we only need the classpath, not the input files)
-    store.materialize_directory(workdir_for_server.clone(), input_files, workunit_store)
-        .and_then(move |_metadata| {
-          let jdk_home_in_workdir = &workdir_for_server.clone().join(".jdk");
-          let jdk_home_in_workdir2 = jdk_home_in_workdir.clone();
-          let jdk_home_in_workdir3 = jdk_home_in_workdir.clone();
-          if jdk_home_in_workdir.exists() {
-            let maybe_existing_jdk = read_link(jdk_home_in_workdir).map_err(|e| format!("{}", e));
-            let maybe_existing_jdk2 = maybe_existing_jdk.clone();
-            if maybe_existing_jdk.is_err() || (maybe_existing_jdk.is_ok() && maybe_existing_jdk.unwrap() != requested_jdk_home) {
-              remove_file(jdk_home_in_workdir2)
-                  .map_err(|err| format!(
-                    "Error removing existing (but incorrect) jdk symlink. We wanted it to point to {:?}, but it pointed to {:?}. {}",
-                    &requested_jdk_home, &maybe_existing_jdk2, err
-                  ))
-                  .and_then(|_| {
-                    symlink(requested_jdk_home, jdk_home_in_workdir3)
-                        .map_err(|err| format!("Error overwriting symlink for local execution in workdir {:?}: {:?}", &workdir_for_server, err))
-                  })
-            } else {
-              debug!("JDK home for Nailgun already exists in {:?}. Using that one.", &workdir_for_server);
-              Ok(())
-            }
-          } else {
-            symlink(requested_jdk_home, jdk_home_in_workdir)
-                .map_err(|err| format!("Error making new symlink for local execution in workdir {:?}: {:?}", &workdir_for_server, err))
-          }
-        })
-        .inspect(move |_| debug!("Materialized directory {:?} before connecting to nailgun server.", &workdir_for_server2))
-        .to_boxed()
-  }
 }
 
 impl super::CommandRunner for CommandRunner {
@@ -302,27 +258,21 @@ impl CapturedWorkdir for CommandRunner {
     let nails_command = self
       .async_semaphore
       .with_acquired(move || {
-        Self::materialize_workdir_for_server(
-          store,
-          workdir_for_this_nailgun.clone(),
-          jdk_home,
-          req.input_files,
-          workunit_store,
-        )
-        .and_then(move |_metadata| {
-          // Connect to a running nailgun.
-          executor.spawn_on_io_pool(futures::future::lazy(move || {
-            nailgun_pool.connect(
-              nailgun_name.clone(),
-              nailgun_req,
-              &workdir_for_this_nailgun1,
-              nailgun_req_digest,
-              build_id,
-            )
-          }))
-        })
-        .map_err(|e| format!("Failed to connect to nailgun! {}", e))
+        // Connect to a running nailgun.
+        executor.spawn_on_io_pool(futures::future::lazy(move || {
+          nailgun_pool.connect(
+            nailgun_name.clone(),
+            nailgun_req,
+            workdir_for_this_nailgun1,
+            nailgun_req_digest,
+            build_id,
+            store,
+            req.input_files.clone(),
+            workunit_store
+          )
+        }))
       })
+      .map_err(|e| format!("Failed to connect to nailgun! {}", e))
       .inspect(move |_| debug!("Connected to nailgun instance {}", &nailgun_name3))
       .and_then(move |nailgun_port| {
         // Run the client request in the nailgun we have active.
